refactor: reorganize DTOs and add generic ApiResponse for consistency
- Moved EventDto from infra.dtos to infra.presentation.dtos
- Updated EventDtoMapper to reference the new package
- Introduced ApiResponse<T> as a generic response wrapper
- Updated EventController to return ResponseEntity<ApiResponse<EventDto>> on creation
- Renamed EventRepositoryGateway to EventGatewayImpl for better naming consistency
