+++ src/main/java/br/com/carloslonghi/eventflow/api/infra/mapper/EventDtoMapper.java
@@ -1,7 +1,7 @@
 package br.com.carloslonghi.eventflow.api.infra.mapper;
 
 import br.com.carloslonghi.eventflow.api.core.domain.Event;
-import br.com.carloslonghi.eventflow.api.infra.dtos.EventDto;
+import br.com.carloslonghi.eventflow.api.infra.presentation.dtos.EventDto;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
